tg_bot/requirements.txt
aiogram==3.20.0
python-dotenv==1.0.0
kafka-python==2.0.2
Django==5.2.1
psycopg2-binary==2.9.9
djangorestframework==3.16.0


django-filter==25.1

# Разработка
